#103: review adjustments

+++ cli/src/main/java/com/devonfw/tools/ide/url/model/file/json/CVE.java
@@ -7,6 +7,6 @@ import com.devonfw.tools.ide.version.VersionRange;
 /**
  * Model to represent a CVE of a tool (inside a "security.json" file).
  */
-public record CVE(String id, float severity, List<VersionRange> versions) {
+public record CVE(String id, double severity, List<VersionRange> versions) {
 
 }
